requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==6.0.4
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-mptt==0.18.0
django-redis==6.0.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
fire==0.7.1
flower==2.0.1
gunicorn==25.3.0
humanize==4.15.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
markdown-it-py==4.0.0
mdurl==0.1.2
openpyxl==3.1.5
packaging==26.0
pillow==12.2.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.20.0
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
rav==0.4.3
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rich==14.3.4
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
termcolor==3.3.0
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
